MC Errors Increase Size (#464)
* changed errors and warnings variable to 32 bits

* reverted error source codes

* changed variable names according to size

* changed sizeof to 32 in comm_task

* switched from 64 to 32 bits

* took out unecessary casting

+++ src/GNR2_Firmware/revA/grn2_ra6t2_ftexboard/FTEX/inc/middleware_motor/ics_ra6t2_pwm_curr_fdbk.h
@@ -127,7 +127,7 @@ void PWMInsulCurrSensorFdbk_SwitchOffPWM(PWMCurrFdbkHandle_t * pHdl);
 	* @param  pHdl: handle of the current instance of the PWMInsulCurrSensorFdbkHandle_t component.
 * @retval Motor control error code: MC_FOC_DURATION if overrun occured, MC_NO_FAULTS otherwise.
   */
-uint16_t PWMInsulCurrSensorFdbk_WriteTIMRegisters(PWMCurrFdbkHandle_t * pHdl);
+uint32_t PWMInsulCurrSensorFdbk_WriteTIMRegisters(PWMCurrFdbkHandle_t * pHdl);
 
 
 /**
@@ -150,7 +150,7 @@ void * PWMInsulCurrSensorFdbk_BRK_IRQHandler(PWMInsulCurrSensorFdbkHandle_t * pH
   * @retval uint16_t It returns MC_BREAK_IN whether an overcurrent has been
   *                  detected since last method call, MC_NO_FAULTS otherwise.
   */
-uint16_t PWMInsulCurrSensorFdbk_IsOverCurrentOccurred(PWMCurrFdbkHandle_t * pHdl);
+uint32_t PWMInsulCurrSensorFdbk_IsOverCurrentOccurred(PWMCurrFdbkHandle_t * pHdl);
 
 
 #ifdef __cplusplus
